fe: Fix perf issue in LibraryFeature.isUniverse and some cleanup
`LibraryFeature.isUniverse()` used about 10% of the DFA time in fzweb due to the
call to an expensive call to `LibraryModule.libraryUniverse()`. Replaced this a
call to `LibraryFeature.universe()` that uses a field.

Then, replaced calls to `libraryUniverse` by `universe` where possible and
replaced the only remaining call to `libraryUniverse` by a private field
`_libraryUniverse`.