dfa: Improve the fixpoint loop: run one more iteration after `lookupDone`
After `lookupDone` was called, some values will be turned into unit type values.
This may have the effect that new clazzes are created if these unit types are
boxed.

So we run one more iteration before we report the results and check that the
result reporting perfomed no more changes to the reached fix point.