dfa: Performance improvement by originally treating calls as unit type calls
Only when a feature has fields that are read, new calls for each context will be
created.

This improves the DFA performance on a HelloWorld somewhat.

This has slight risk that a several data flow from a different call site may
flow into a call site provided until it was found that that call site is not a
unit type instance. A fix would be to throw away the unit type call altogether
once it was found that it is not on a unit type.