jvm: A `staticAccess` in a stub was falsely optimized as a tail call
Stubs are used to implement dynamic binging using Java's interface calls: A call
is replaced by a interface call to a stub which is implemented by all actual
implementations of the call.

The stub itself then performs a static call to the actual feature.

The problem was that tail call optimization might be done by `staticAccess` if
the target is the same as the current clazz.  For the code generation of a stub,
however, the `site` that was provided was the first call site that causes the
generation of the `stub`. If this happened to be in the same clazz as the called
clazz, tail call optimization inserted a goto in the stub to the start of the
method created for the called clazz, usually resulting in a class file verifier
error.

This could be triggered by running tests/transducers with JVM backend:

    build/tests/transducers 16:06:47 > dev_flang_fuir_analysis_dfa_DFA_SITE_SENSITIVE=false make jvm
    ../check_simple_example_jvm.sh "../../bin/fz " transducertest.fz || exit 1
    RUN transducertest.fz 0a1,23
    >
    > error 1: java.lang.VerifyError: Expecting a stackmap frame at branch target 8
    > Exception Details:
    >   Location:
    >     fzC__L1230_transducer__lter__fun__1call__fun.fzD_1285_call(LfzI_Sequence_i32;I)LfzI_Sequence_i32; @6: goto
    >   Reason:
    >     Expected stackmap frame at this location.
    >   Bytecode:
    >     0000000: 2a4b 2b4c 1c3d a700 02
    >   Stackmap Table:
    >
    > 	at fzC__L1226_transducer__test_human__1filter__fun__1call.fzRoutine($MODULE/transducer.fz:66)
    > 	at fzC_1transducer__Sequence_i32__i32__i32__1call.fzRoutine($MODULE/transducer.fz:51)
    > 	at fzC_3interval_i32__1into_i32.fzRoutine($MODULE/Sequence.fz:540)
    > 	at fzC_transducertest.fzRoutine(--CURDIR--/transducertest.fz:39)
    > 	at fzC_universe.fz_run(--builtin--)
    > 	at dev.flang.be.jvm.runtime.FuzionThread.lambda$new$1(FuzionThread.java:120)
    > 	at dev.flang.util.Errors.runAndExit(Errors.java:897)
    > 	at dev.flang.be.jvm.runtime.FuzionThread.lambda$new$2(FuzionThread.java:132)
    > 	at java.base/java.lang.Thread.run(Thread.java:1583)
    >
    > *** fatal errors encountered, stopping.
    > one error.
    *** FAILED err on transducertest.fz
    make: *** [../simple.mk:64: jvm] Fehler 1