analysis: Added version of AbstractInterpreter without RESULT
This provides a slight speedup for uses of the AbstractInterpreter that use a
unit type as RESULT while Java does not really provide a unit type.

When buildig `fzweb` into a JAR, I gain about 1.7% (33.14s instead of 33.70s
elapsed time).

To avoid the risk that duplicated code will eventually get out of sync, I use a
patch to generate the new variant of AbstractInterpreter.  Explanation on how to
update the patch is given in the source code.