ast/fe: Attempt to fix #3731
First, this removes the special handling for `ref` types in
`Call.adjustThisTypeForTarget`.

Then, it calls `replace_this_type_by_actual_outer` using `_target.type()` and
not `inner`, which is a workaround to find `Any.this` in the target and not in
`me`.

Finally, it disables a cyclic field error that occurs for some unknown reason.