tests: in `negative.fz` use `$(FILE)` instead of `$(NAME)`
This makes it clear with source file is loaded and avoides searching for source files.

In particular, this also permits executable code in the universe of that source
file, otherwise, it would be ambiguous when to execute that code.

This revealed an error in `reg_issue4273` where the negative tests were ignored
due to a type in the comments resulting in the lines that must fail not being
found by `grep`.