ast: avoid repeated immediate function call if resolveTypes called repeatedly
The following gives the smalles example I found that caused this problem:

  b (t Lazy nil) => b (b t)

The problem is that the call to `t` is resolved repeatedly hence gets converted
in `t.call` and `t.call.call` repeatedly resulting in require-conditions to fail
in `fz`.

This patch uses a flag `_pushImplicitImmediateCall` to avoid repeated
application of immediate call syntax sugar.