dfa: fix type of valueset of original of TaggedValue
A TaggedValue may have a whole set of values as original values. When joining
two TaggedValue instances of the same tag, the _clazz of the joined original
values was falses set the the type of the tagged value, it has to be the type of
the original value.

Also marked some fields final since that help me debugging by making sure these
are immutable.