lib: add `container/expaning_array` and `container/abstract_array`
`expanding_array` provides an efficient dynamically growing array that does not
require a `mutate` effect since it does not mutate any existing instances.  But
it attempts to share the internal array data with earlier instances, so it
should achieve performance similar to Java's `ArrayList`.

Added `abstract_array` to reduce code duplication in `array` and
`expanding_array`.