ai: Handle `match` that never returns in AbstractInterpreter, fix #4662
Once the DFA has found that a `match` would never return, stop
interpreting. Since the result of a match by the backend could only be a unit
type anyway, this takes the burdern of return `null` in case of an effective
`void` result form the backends.

Only uses of the `AbstractInterpreter` that directly return a value, i.e., the
DFA and the interpreter backend, still use `null` to signal no return (DFA), or
just never actually return (interpreter).