ast: Fix partial application for infix operators, fix #4687
Partial application was not done for infix operator if call resultion did not
produce a pending error.  Added check for `isPartialInfix` in
`ParsedCall.propagateExpectedTypeForPartial` to perform partial application in
this case.

Also did some cleanup: Call ot `newNameForPartial` in
`Call.partiallyApplicableAlternative` was not needed, so I removed this method
altogether.

Added `isPartialInfix` as an alternative to `newNameForPartial` and added the
code to actual create the new name as inline code to the only remaining place
that needs the name.