lib: Workaround for performance issue in DFA when running `tests/process`
This performance issue is likely related #4634, using `expanding_array.expand`
results in additional environment contexts that result in a surprising explosion
of call contexts.

Also, added special handling for concat with an empty Sequence and added
comments.

A bit tricky: `expanding_array.concat` was marked `fixed`, which is only needed
to simplify the code in the final `else` clause's loop and avoid intermediate
boxing.