build: Enable building `fuzion.ebnf` in a dir different than the source dir
It should be possible to build `fuzion` from any dir using `make -f
<fuzion-dir>/Makefile`. This did not work for `fuzion.ebnf` and, what is worse,
the error was silently ignored with an empty `fuzion.ebnf` file.

Now, `ebnf.fz` expects the source files that contain ebnf data as
arguments. Furthermore, it produces errors in case these arguments are missing
or in case the files do not exist.