dfa: Do not call `result()`in `Call.toString()`.
`result()` has all sorts of side effects since it tries to determine the result
by code analysis.  Since `toString()` is used for debugging, this will change
the state of the DFA in a way that makes this debugging harder.

So, I prefer printing no details on the results and not changing the state in
`toString()` to not mess up with what I am debugging.