ast: Fix checking of actual type parameters, fix #5001
If a type constraint uses type parameters itself and those are defined by the
target of a call, we need to first determine the actual types of the constraints
in this call before we can check the validity of the actual type parameters
provided by the call.

Added a check that `AbstractType.genericsAssignable` does not deal with open
type parameters unless in a cotype (which should be fixed later).

Also added a reference to #5002 in the correspdonging comment.