lib, doc: Workaround to improve DFA performance: Move local instances to `unit`
This is a workaround for #5036 in the base library and in `bin/ebnf.fz` to use
`unit.lm` instead of a local `lm` instance to simplify the DFA analysis.

Without this, the values of the local instances get duplicated for each outer
value, resulting in an explosion of the overhead of the DFA analysis.

This patch reduces the runtime of `ebnf.fz` without argument by more than 30%
(from 24s down to 15s on my machine), the biggest part is the change in `ebnf.fz` itself, the others add about 0.5-1sec.