fe: supresss subsequent error on argument count in #5114
With the null pointer fixed, the example from #5114 results in two additional
errors on the argument count of the redefintion of Function.call not matching,
which are just caused because `handDown` could did not operate properly.

Now, `handDown` uses a special empty array as result value in case it failed due
to previous errors.

Also made the code in `Function` more robust against subsequent errors in case
the argument count of a lambda does not match.

Fixed a missing ".\n" in error for argument count mismatch in redefinition.