parser: cleanup: Remove `mayUseCommas` argument for rules like `expr`
Instead, this adds `Lexer.endAtComma` similar to the existing `endAtColon`,
`endAtBar` to stop the parser that the next `,`. This simplifies the parser at
the expense of a slightly more complex lexer.

Also added some missing comments to the EBNF grammar where rules are parsed only
as long as no `,` or `:` is encountered.

Fixed reference to #3798 (that falsely pointed to #3978).