dfa: Two phase DFA
Run DFA in two phases: First, to collect calls in `CallGroups` with common call
destination, call site and call target to find all environments required by this
call group.

Second, to run the analysis tracing effect environments as well, but limiting
this to the environments that are actually required as found in the first phase
for the group each call belongs to.