ast: fix handling of open type parameters when resolving types
This fixes #5436 by adding support for replacing open type parameters by a list
of actual types.

Since this requires the use of `flatMap` instead of `map` since the numer of
type parameters may change, and `List.flatMap` always allocated a new array even
if nothing changed this first failed since the code relies on list of generics
to stay equal if there is no change.  For this reason, and for the probable
performance improvement, `List.flatMap` was changed to create a new `List` on
demand only.

A new helper `List.take(n)` was added to implement `List.flatMap()`.

`AbstractType.applyTypeParsLocally` no longer applies actual generics to a this
type, which was the cause of the third failure in #5436.

Added a test with variants of the examples from #5436.