parser: unify handling of `()` and `{}` with respect to indentation, fix #4425
Now, `{`  and `(` can be used largely interchangably around code block, e.g., we can create a block as follows

    say (if 3 = 4 then
           say "strange"
         else
           say "this should run 1"
         "result from code block 1")

which is equivalent to the same code using `{`/`}`:

    say {if 3 = 4 then
           say "strange"
         else
           say "this should run 2"
         "result from code block 2"}

Simplified the grammar rules for `klammer`, `tuple` and `lambda`.

There are still some minor syntactical differences if one a single
`operatorExpr` is put in `(`/`)` vs. `{`/`}`, in particular

    _ := l.zip m (a,b -> unit)  # with {}, would be parsed as declaration of `a` and `b` and not lambda
    _ := ("bla"
           + "blub")            # with {}, causes indentation error
    _ := (a).this               # with {}, causes qualifier expected for '.this' expression.

I will leave these for later to be fixed.

Also, there are still differences when parsing actual arguments in parentheses
vs. a tuple with respect to the handling of commas: `(x,y->unit,z)` would be
parsed as three actual arguments (`x`, `y->unit` and `z`), but as a two-tuple
(cosisting of `x, y->unit` and `z`).

Add a regression test for #4425.