ast: set inferred argument field types before checking argument types in call
This ensures that failure to infer the argument type due to incompatible actual
arguments as in the exampe from #5560

    f(v) => say v
    f "a"
    f 1

will be reported first and the type of the argument will be set to t_ERROR such
that no subsequent error will be shown for the actual arguments passed to the
call.

This required explicitly checking the called feature's types first and making
Feature.checkTypes() re-entrant.

The stricter checking required a minor change in a loop in `Sieve_Cache` that
can no longer use auto-unwrap.

fix #5560.