ast: Support for types that depend on outer type of type parameter constraint
A feature `f` with a type parameter `T` and a constraint `a.b.c` as follows

    f(T type : a.b.c) =>

may be used to produce types that depend on the actual types corresponding to
`a.b` or just `a` as passed to `f`.  This was not supported so far resulting in
`fz` crashing.

In case we have

    a is
      b is
        c is
          j is
          get_i => i
          get_j => j
        i is

Before this commit, the result type of `get_i`/`get_j` was `a.b.this.i` and
`a.b.c.this.j`. When called within `f`, these only the one for `get_j` was
replaced by `T.j`, while the one for `get_i` remained `a.b.this.i`, which did not get replaced during monomorphization.

Now, the types used in `f` will be `T.i` and `T.j`, respectively, and
monomorphization will replace them with the actual clazz used for `T` or the
corresponding outer clazz.

fix #5569