ast/fuir: Improvements suggested by review comments
Added missing call to `resolve` for `constraint0` which was the reason I had to
add `UnresolvedType.applayTypeParsLocally`. The latter could now be removed.

Removed additional `child` iteration in `Clazz` that is actually not needed any
more since the changes to `AbstractType` are sufficient.

Added helper method `AbstractType.replacedThisType`.

Fixed indentation.