fe/me/base: Support to call fields whose type is an open type parameter
The result of such a call is an instance of a child of `TypedSequence` that
permits the application of a  `TypedApplicator` to all the elements.

Use this mechanism to implement `tuple.as_string`

This magic happens mostly in the monomorphization phase that replaces an
intrinsic used used in `TypedSequence.foldf` by the actual code required to apply the applicator to all fields created for the actual type parameters.

This is still work in progress, needs cleanup and testing.