base: Rename `list.scan` as `list.scan_list`, change result of `Sequence.scan` to `Sequence`
`Sequence.scan` was the only remaining feature in `Sequence` that returned a
`list` and not a `Sequence`, so this is fixed now.

Renaming the versions in `list` that produce a `list` is for consistency and to
fix the following confusion:

I stumbled into this since I was suprised that adding `as_list` in the following
example caused an error:

     > ./build/bin/fz -e "[1,2,3].scan 100 (+) |> say"
    [100, 101, 103, 106]
     > ./build/bin/fz -e "[1,2,3].as_list.scan 100 (+) |> say"

    command line:1:26: error 1: No type information can be inferred from a lambda expression
    [1,2,3].as_list.scan 100 (+) |> say

The reason was that `list` adds another variant of `scan` with 2 arguments.