examples: cleanup in jitter
Extract start/end and header output outside of state effect for next_release.

reuse `start` value as initial value for `next_release`.

Replace type for state effect by `_`, type inference works here.

In `if .. else`, there is no need for the `else` if the code in `if` results in
`void`, so removed the `else`.