ast: fix inheritance from feature with open type parameters
The problem was that the actual number of type parameters could change and even
become smaller if open type parameters are involved.

This patch also simplifies the code by first using the type parameters from the
original (non-cotype) feature and replacing them once the type parameters of that feature have been inferred.

Fix #5821 and add disabled tests for #5835.