ast: Fix check condition failure when comparing tuples
The problem was that after a `fum` file is loaded, the magic done using
`FormalGenerics.AsActuals` no longer works.  Since this was only intended as an
optimization, it got overlooked that the generic code to replace an open type
parameter did not handle open type parameters in cotype correctly.

This patch adds a new method `matchingTypeParameter` that extracts the code
relevant for checking if a type parameter comes from a given feature even if it
is a clone in the cotype.  This code is now used in in the methods
`applyTypeParsLocally` (which does not handle open type parameters) and
`applyTypePars()` with List arguments, that handles open type parameters as
well.

Also added a regression with the original failing code and more tests using
ordered maps from tuples to i32 and vice versa.