base: add `Any.dynamic_apply` and `Typed_Function`
This allows, e.g., calculating the sum of all `i32` and `f64` elements in a `Sequence Any`:

    sum(values Sequence Any)
    =>
      get_f64 : Typed_Function f64 is
        public redef call(T type, v T) f64
        =>
          if      T : i32 then v.as_f64
          else if T : f64 then v.val
                          else compile_time_panic

      values.map (.dynamic_apply get_f64)
            .foldf 0.0 (+)

    say <| sum [3.14, 32, 16.0, 8]

`dynamic_apply` can be used similar to `dynamic_type` to reflect on the dynamic
type, but it also provides the value using that type such that we can write
write code that gets specialized for specific dynamic types.

Added a new feature `Typed_Function` for this.

Using `dynamic_apply` would be nicer if `Typed_Function` would be supported by
our lambda syntax sugar #5892.

`Any.dynamic_apply` should be changed to use an explicit type parameter for
readability, but this is not possible due to #5894.

The argument order in `Typed_Function` should IMHO be changed, but cannot
because of #5895.