ast/fuir: Fix using open type parameters in inherted feature
The problem was twofold:

First, `AbstractFeature.handDownInheritance` replaced the open generic type
parameters with whatever type parameters where given in the inheritance call.

Second, if there was inheritance, `Clazz.replaceOpen` always tried to replace
open type parameters by the inheritance call even if they are not inherited.

In both cases, the additional check that the called feature in the inheritance
call had to be the declaring feature of the open type parameter.

Added a regression test.

fix #5941