Makefile: reorder include mod_java.mk after all target
When just typing `make`, the default target is the first target listed
in the Makefile. Due to the include, this was a `__marker_for_make__`
target. If we move the include, the default target is `all` again.

The new position needs to be chosen carefully. If any dependency is set
on variables declared in mod_java.mk before the include, the build will
fail.